python/requirements.txt
cryptography>=41.0.0
python-dotenv>=1.0.0
flask>=3.0.0 